+++ src/utils/getRandomNumber.js
@@ -1,0 +1,3 @@
+export function getRandomNumber(start, end) {
+  return Math.floor(Math.random() * (end - start + 1)) + start;
+}
